Avoid overflow in polynomial & increase tolerance
Since pi/2 - acos is in the range [0, pi/2], the table result sometimes
becomes larger than 1. The current implementation (not intentionally)
use this to keep precision. But if we share polynomial eval module with
other math funcs, we cannot use such a hack. The (possibly only) way to
keep precision is to allow 3 extrabits, instead of 2.