WIP: add pow2sim
to see what happens if we use normal dxW for dx in polynomial.
The pow2 impl in math/exponential uses wider dx for the sake of
accuracy, but in mathfunc/* we use fixed dx in all the math funcs.
To save the bits in polynomial, we need to check how accurate it can be.