Reorder multiplication for the sake of accuracy
If we first multiply x.man and 1/ln2 and round it to prec = manW+extra,
the result would have large error (like, 5LUPs). If we first multiply
x and (1-x/2) and then multiply it with 1/ln2, the error dissapears.
But, from the viewpoint of speed, (x * 1/ln2) * (1 - x/2) is faster than
(x * (1-x/2)) * 1/ln2 because of the data dependency...