Widen the address bits of table for ex==0
The accuracy of log2 function decreases around x=1 because the integer
part of z is zero and log(1.man) table has limited precision. To
workaround this, we added a table specialized for the case of x = [1,2).
But the accuracy didn't increase so much. It is suspected that the shape
of the function decreases the accuracy. Here, to gain more accuracy, we
shifted the result using the number of zero bits in the mantissa of x.
So the table tries to interpolate log2(1.0+x) * pow2(xex-2). Since the
function interpolated changes the value drastically every time the zero
bit appears at the MSB of x, the function shape becomes jaggy. This may
inhibits accurate interpolation. To avoid this effect, we increased the
width of address bit. The accuracy increased in return for the larger
table size. Also, the simulator becomes much slower.