Add full range cos(x) to mathfunc
The first conversion x -> x/pi requires extremely wide Pi to achieve the
error within the last 2 bits compared to math.cos(x:Double).
Also, this implementation uses 5th order term in the Taylor expansion
around x=0 to reduce the required number of tables. This requires 5
steps in the non-Taylor path that is 1 step more than that in the
polynomial path.
Even so, this function still have 2-bit error. The reason is that the
implementation re-uses table of sin. This means cos(x) around x=0 is
calculated by converting sin(pi-x). This introduces one bit error that
is anavoidable even if we use wider polynomial or higher order Taylor
expansion. The only way to reduce the error is to increase the number
of tables by implementing cosine table independent from sin table.