Merge equivalent ShiftRegisters into one Reg
So far, we only used register that value is up-to-date to the timing
when some module returns a value. But in many cases, we have a register
in between module input/outputs and we needed one ShiftRegister there.
That looks clearer maybe, because module-specific ports are also delayed
by ShiftRegister, but could make FIR redundant because ShiftRegisters
of shared values, like decomposed x and selection signal, are equivalent.