Remove () from 0-arg getters with no side effect
Auto-application to `()` is deprecated.
It seems that, customarily, people call zero-argument function with
explicit () if the function has side effect, i.e. the function cannot be
replaced with a member variable. If a function does not have any side
effect, we can replace it by a variable access. In such cases, people
calls the function without any paren.
Here we remove `()`s from getters with no side effects.