Test sincos only module (WIP: BF16)
As written in the comment, sincos modules has a problem with a FP having
narrow mantissa (e.g. BF16). The taylor expansion uses very high-order
term, like 5th order term. This is enough for BF16 to fill the mantissa
everywhere. So the table does nothing and causes unbalance. We need to
find a smart way to calculate taylorThreshold or to provide a way to set
the taylor expansion order manually.