Change sincos table approx to gain more precision
since sin(pix)/x is in (2, pi] when x is in [0, 1/2), it fits to
the range that corresponds to ex == 1. so we can remove the "hidden
bit". by changing table approx function from sin(pi*x)/4x to
sin(pi*x)/2x - 1, we can have one extra bit at the LSB and gain a bit
more precision.